+++ CHANGELOG.md
@@ -18,3 +18,4 @@
 - Rinominati elementi TrickEntry, TrickEntry[] e TrickEntry[][] in TrickEntry, Trick, and CompletedTricks
 - refactored ComputerPlayer externalizing methdos. Enabled strict compile mode, an converted ComputerPlayer, GameState, and some other minor JS in TS.
 - Aggiunto interruttore audio
+- Aggiunti Confetti con la libreria confetti.js (https://confettijs.org/) e audio nel GameOver